fix: APIError localizedDescription
