+++ Source/API/APIResult.swift
@@ -8,26 +8,6 @@
 
 import Foundation
 
-public struct APIError: Codable, LocalizedError {
-    let code: Int
-    let message: String
-
-    var localizedDescription: String {
-        return message
-    }
-}
-
-extension APIError {
-    static let genericErrorCode = -1
-    static let invalidParameters = APIError(code: genericErrorCode, message: "Invalid parameters")
-    static let emptyResponse = APIError(code: genericErrorCode, message: "Empty response")
-    static let nullResult = APIError(code: genericErrorCode, message: "Null result")
-    static let unmatchedId = APIError(code: genericErrorCode, message: "Unmatched id")
-    static func genericError(_ message: String) -> APIError {
-        return APIError(code: genericErrorCode, message: message)
-    }
-}
-
 /// JSON API Response
 struct APIResult<R: Codable>: Codable {
     let jsonrpc: String
